Server.py
This is the server socket. Depending on the command, it receives and sends messages to users (privately or in group chat(network)). I've used a randomly generated sequence of numbers for certain commands like Broadcast or Private Message to distinguish specific uses. (Maybe there are some methods that are actually useless, but I'm afraid to delete those because they might break my whole code