Create Gruppenchat.py
This is the client socket. The network class represents the group rooms and has properties for how many clients can use them and how many subnets, i.e. subgroup rooms, it has (I didn't take the subnets into account in the socket programming because this is unnecessary for my learning experience and just more work might). There are other classes, such as Message, Groupchat. These are important to save the messages and display them on the GUI. (Their also might be some useless code...)